GUM-24: app-color file renamed
