+++ src/common/enums/ui/app-color.ts
