+++ src/common/enums/ui/index.ts
@@ -1,4 +1,4 @@
-export * from './app-colors';
+export * from './app-color';
 export * from './button-mode';
 export * from './heading-level';
 export * from './text-appearance';
